moved action menu to zone card and created edit zone function
